Working click movement
