+++ .mono/metadata/scripts_metadata.editor_player
@@ -1,1 +1,1 @@
-{"res://PlanetData.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"PlanetData","nested":false}},"res://PlanetGen.cs":{"modified_time":"1675168014","class":{"namespace":"","class_name":"PlanetGen","nested":false}},"res://PlayerCam.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"PlayerCam","nested":false}},"res://Mars.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"Mars","nested":false}},"res://Planet.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"Planet","nested":false}},"res://HTTPRequestDemo.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"HTTPRequestDemo","nested":false}},"res://Player.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"Player","nested":false}},"res://Coordinates.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"Coordinates","nested":false}},"res://MarsMesh.cs":{"modified_time":"1675168014","class":{"namespace":"","class_name":"MarsMesh","nested":false}},"res://PlanetMeshFace.cs":{"modified_time":"1675775437","class":{"namespace":"","class_name":"PlanetMeshFace","nested":false}},"res://BetterPlayer.cs":{"modified_time":"1675778538","class":{"namespace":"","class_name":"BetterPlayer","nested":false}}}
+{"res://Coordinates.cs":{"modified_time":"1670710473","class":{"namespace":"","class_name":"Coordinates","nested":false}},"res://HTTPRequestDemo.cs":{"modified_time":"1670885875","class":{"namespace":"","class_name":"HTTPRequestDemo","nested":false}},"res://Planet.cs":{"modified_time":"1675191047","class":{"namespace":"","class_name":"Planet","nested":false}},"res://PlanetData.cs":{"modified_time":"1675946928","class":{"namespace":"","class_name":"PlanetData","nested":false}},"res://PlanetGen.cs":{"modified_time":"1675188120","class":{"namespace":"","class_name":"PlanetGen","nested":false}},"res://PlanetMeshFace.cs":{"modified_time":"1675946928","class":{"namespace":"","class_name":"PlanetMeshFace","nested":false}},"res://PlayerCam.cs":{"modified_time":"1675900036","class":{"namespace":"","class_name":"PlayerCam","nested":false}},"res://Mars.cs":{"modified_time":"1675949080","class":{"namespace":"","class_name":"Mars","nested":false}},"res://MarsMesh.cs":{"modified_time":"1675949359","class":{"namespace":"","class_name":"MarsMesh","nested":false}},"res://MainCam.cs":{"modified_time":"1675949675","class":{"namespace":"","class_name":"MainCam","nested":false}},"res://Player.cs":{"modified_time":"1675949677","class":{"namespace":"","class_name":"Player","nested":false}},"res://BetterPlayer.cs":{"modified_time":"1675950970","class":{"namespace":"","class_name":"BetterPlayer","nested":false}}}
