+++ Mars.cs
@@ -3,95 +3,95 @@ using System;
 
 public class Mars : StaticBody
 {
-    // Declare member variables here. Examples:
-    // private int a = 2;
-    // private string b = "text";
-    Vector2 MouseDelta;
-    Vector2 NextMousePos, PrevMousePos;
-    bool MouseEntered = false;
-    bool Rotating = false;
-    public float rotationSpeed = 0.5f;
-    bool RotatingX, RotatingY, RotatingZ;
-    Tween Snapback;
+    // // Declare member variables here. Examples:
+    // // private int a = 2;
+    // // private string b = "text";
+    // Vector2 MouseDelta;
+    // Vector2 NextMousePos, PrevMousePos;
+    // bool MouseEntered = false;
+    // bool Rotating = false;
+    // public float rotationSpeed = 0.5f;
+    // bool RotatingX, RotatingY, RotatingZ;
+    // Tween Snapback;
 
-    public override void _Ready()
-    {
-        RotatingX = false;
-        RotatingY = false;
-        RotatingZ = false;
-        Snapback = GetNode<Tween>("Snapback");
-    }
+    // public override void _Ready()
+    // {
+    //     RotatingX = false;
+    //     RotatingY = false;
+    //     RotatingZ = false;
+    //     Snapback = GetNode<Tween>("Snapback");
+    // }
 
-    public override void _Process(float delta)
-    {
-        LocationLock();
+    // public override void _Process(float delta)
+    // {
+    //     LocationLock();
 
-        if (Input.IsActionJustPressed("XRot"))
-        {
-            RotatingX = true;
-            RotatingY = false;
-            RotatingZ = false;
-        }
-        if (Input.IsActionJustPressed("YRot"))
-        {
-            RotatingY = true;
-            RotatingX = false;
-            RotatingZ = false;
-        }
-        if (Input.IsActionJustPressed("ZRot"))
-        {
-            RotatingZ = true;
-            RotatingX = false;
-            RotatingY = false;
-        }
-
-
-
-        if (Input.IsActionJustPressed("mousepress") && MouseEntered)
-        {
-            PrevMousePos = GetViewport().GetMousePosition();
-            Rotating = true;
-        }
-        if (Input.IsActionJustReleased("mousepress"))
-        {
-            Rotating = false;
-        }
-        if (Input.IsActionJustPressed("ui_escape"))
-        {
-            Snapback.InterpolateProperty(this, "rotation_degrees", RotationDegrees, new Vector3(0, 0, 0), 0.5f, Tween.TransitionType.Elastic, Tween.EaseType.InOut);
-            Snapback.Start();
-        }
-
-        if (Rotating)
-        {
-            NextMousePos = GetViewport().GetMousePosition();
-
-            if (RotatingX) RotateX((NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
-            if (RotatingY) RotateY((NextMousePos.x - PrevMousePos.x) * rotationSpeed * delta);
-            if (RotatingZ) RotateZ(-(NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
-            // RotateX((NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
-            // RotateY((NextMousePos.x - PrevMousePos.x) * rotationSpeed * delta);
-            // RotateZ(-(NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
-
-            PrevMousePos = NextMousePos;
-        }
+    //     if (Input.IsActionJustPressed("XRot"))
+    //     {
+    //         RotatingX = true;
+    //         RotatingY = false;
+    //         RotatingZ = false;
+    //     }
+    //     if (Input.IsActionJustPressed("YRot"))
+    //     {
+    //         RotatingY = true;
+    //         RotatingX = false;
+    //         RotatingZ = false;
+    //     }
+    //     if (Input.IsActionJustPressed("ZRot"))
+    //     {
+    //         RotatingZ = true;
+    //         RotatingX = false;
+    //         RotatingY = false;
+    //     }
 
 
-    }
 
-    private void _on_Area_mouse_entered()
-    {
-        MouseEntered = true;
-    }
-    private void _on_Area_mouse_exited()
-    {
-        MouseEntered = false;
-    }
+    //     if (Input.IsActionJustPressed("mousepress") && MouseEntered)
+    //     {
+    //         PrevMousePos = GetViewport().GetMousePosition();
+    //         Rotating = true;
+    //     }
+    //     if (Input.IsActionJustReleased("mousepress"))
+    //     {
+    //         Rotating = false;
+    //     }
+    //     if (Input.IsActionJustPressed("ui_escape"))
+    //     {
+    //         Snapback.InterpolateProperty(this, "rotation_degrees", RotationDegrees, new Vector3(0, 0, 0), 0.5f, Tween.TransitionType.Elastic, Tween.EaseType.InOut);
+    //         Snapback.Start();
+    //     }
 
-    private void LocationLock()
-    {
-        Translation = new Vector3(0, 0, 0);
-    }
+    //     if (Rotating)
+    //     {
+    //         NextMousePos = GetViewport().GetMousePosition();
+
+    //         if (RotatingX) RotateX((NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
+    //         if (RotatingY) RotateY((NextMousePos.x - PrevMousePos.x) * rotationSpeed * delta);
+    //         if (RotatingZ) RotateZ(-(NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
+    //         // RotateX((NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
+    //         // RotateY((NextMousePos.x - PrevMousePos.x) * rotationSpeed * delta);
+    //         // RotateZ(-(NextMousePos.y - PrevMousePos.y) * rotationSpeed * delta);
+
+    //         PrevMousePos = NextMousePos;
+    //     }
+
+
+    // }
+
+    // private void _on_Area_mouse_entered()
+    // {
+    //     MouseEntered = true;
+    // }
+    // private void _on_Area_mouse_exited()
+    // {
+    //     MouseEntered = false;
+    // }
+
+    // private void LocationLock()
+    // {
+    //     Translation = new Vector3(0, 0, 0);
+    // }
 
 
 }
