+++ Planet.tscn
@@ -1,4 +1,4 @@
-[gd_scene load_steps=39 format=2]
+[gd_scene load_steps=40 format=2]
 
 [ext_resource path="res://textures/temp_textures/5672_mars_6k_color(1).jpg" type="Texture" id=1]
 [ext_resource path="res://PlanetGen.cs" type="Script" id=2]
@@ -10,6 +10,7 @@
 [ext_resource path="res://PlanetGen.tscn" type="PackedScene" id=8]
 [ext_resource path="res://assets/perlinnose.png" type="Texture" id=9]
 [ext_resource path="res://BetterPlayer.tscn" type="PackedScene" id=10]
+[ext_resource path="res://MainCam.cs" type="Script" id=11]
 
 [sub_resource type="PhysicsMaterial" id=12]
 friction = 0.2
@@ -284,19 +285,27 @@ light_energy = 0.1
 [node name="CollisionShape" type="CollisionShape" parent="Mars"]
 shape = SubResource( 11 )
 
-[node name="Snapback" type="Tween" parent="Mars"]
-
 [node name="MeshInstance" type="MeshInstance" parent="Mars"]
 mesh = SubResource( 13 )
 material/0 = SubResource( 17 )
 
-[node name="MainCam" type="Camera" parent="."]
+[node name="Player" parent="Mars" instance=ExtResource( 10 )]
+transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 3.05463 )
+
+[node name="CameraPivot" type="Spatial" parent="."]
+script = ExtResource( 11 )
+
+[node name="MainCam" type="Camera" parent="CameraPivot"]
 transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 7.446 )
 current = true
+near = 0.01
+far = 8192.0
 
-[node name="OmniLight" type="DirectionalLight" parent="MainCam"]
+[node name="OmniLight" type="DirectionalLight" parent="CameraPivot/MainCam"]
 light_color = Color( 1, 0.854902, 0.768627, 1 )
 light_energy = 0.61
+
+[node name="Snapback" type="Tween" parent="CameraPivot/MainCam"]
 
 [node name="PlanetGen" parent="." instance=ExtResource( 8 )]
 visible = false
@@ -311,11 +320,8 @@ shape = SubResource( 35 )
 
 [node name="CanvasLayer" parent="." instance=ExtResource( 6 )]
 
-[node name="Player" parent="." instance=ExtResource( 10 )]
-transform = Transform( 1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0, 3.05463 )
-
 [connection signal="area_entered" from="Mars/CSGSphere/Area" to="Mars/CSGSphere" method="_on_Area_area_entered"]
 [connection signal="input_event" from="Mars/CSGSphere/Area" to="Mars/CSGSphere" method="_on_Area_input_event"]
-[connection signal="input_event" from="Mars/CSGSphere/Area" to="Player" method="_on_Area_input_event"]
-[connection signal="mouse_entered" from="Mars/CSGSphere/Area" to="Mars" method="_on_Area_mouse_entered"]
-[connection signal="mouse_exited" from="Mars/CSGSphere/Area" to="Mars" method="_on_Area_mouse_exited"]
+[connection signal="input_event" from="Mars/CSGSphere/Area" to="Mars/Player" method="_on_Area_input_event"]
+[connection signal="mouse_entered" from="Mars/CSGSphere/Area" to="CameraPivot" method="_on_Area_mouse_entered"]
+[connection signal="mouse_exited" from="Mars/CSGSphere/Area" to="CameraPivot" method="_on_Area_mouse_exited"]
